refactor: restructure file locations
