--- backend/BlogPost.Api/BlogPost.Application/Contracts/Post/DeletePostRequest.cs
@@ -1,3 +1,0 @@
-namespace BlogPost.Application.Contracts.Post;
-
-public record DeletePostRequest(int Id);
