--- backend/BlogPost.Api/BlogPost.Application/Contracts/Tag/DeleteTagRequest.cs
@@ -1,3 +1,0 @@
-namespace BlogPost.Application.Contracts.Tag;
-
-public record DeleteTagRequest(int Id);
