+++ backend/BlogPost.Api/BlogPost.Application/Posts/Queries/GetPostById/DetailedPostResponse.cs
@@ -1,11 +1,12 @@
-
 namespace BlogPost.Application.Contracts.Post;
 
-public record PostResponse(
+public record DetailedPostResponse(
     int Id,
     string Title,
     string Text,
     DateOnly PostDate,
+    DateTime LastEdit,
     string ImageUrl,
-    IEnumerable<Domain.Entities.Tag> Tags
+    string MarkDownFileLink,
+    IEnumerable<Domain.Tags.Tag> Tags
 );
