+++ backend/BlogPost.Api/BlogPost.Application/Tags/Common/TagResponse.cs
@@ -1,3 +1,3 @@
-namespace BlogPost.Application.Contracts.Tag;
+namespace BlogPost.Application.Tags.Common;
 
 public record TagResponse(int Id, string Name, string Color);
