+++ backend/BlogPost.Api/BlogPost.Domain/Posts/PostErrors.cs
@@ -5,5 +5,5 @@ namespace BlogPost.Domain.Exceptions;
 public static class PostErrors
 {
     public static Error ValidationError(string message) => Error.Validation("Posts.ValidationError", message);
-    public static Error NotFound(int id) => Error.NotFound("Posts.NotFound", $"Post with id {id} was not found.");
+    public static Error NotFoundById(int id) => Error.NotFound("Posts.NotFound", $"Post with id {id} was not found.");
 }
